% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbm_sim.R
\name{FBM}
\alias{FBM}
\title{Fractional Brownian Motion}
\usage{
FBM(N, H, t = 1, method = "davies-harte")
}
\arguments{
\item{N}{Number of simulated points}

\item{H}{Hurst parameter}

\item{t}{Length of the simulated series}

\item{method}{The simulation method to be used}
}
\value{
A single discrete path of fractional Brownian motion by simulating fractional Gaussian noise (see \code{\link{FGN}}).
}
\description{
Returns a discrete simulated realization of fractional Brownian motion of size N.
}
\examples{
\dontrun{
FBM(1000, 0.25, 1.0, 'davies-harte')
}
\dontrun{
FBM(1000, 0.25, 1.0, 'hosking')
}
\dontrun{
FBM(1000, 0.25, 1.0, 'cholesky')
}
}
