% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgn_sim.R
\name{FGN}
\alias{FGN}
\title{Fractional Gaussian Noise}
\usage{
FGN(N, H, method = "davies-harte")
}
\arguments{
\item{N}{Number of simulated points}

\item{H}{Hurst parameter}

\item{method}{The simulation method to be used}
}
\value{
A sample of fractional Gaussian noise of size N
}
\description{
Returns a sample of fractional Gaussian noise of size N
}
